feat(plasma-temple): Empty cart mobile version
